tests/wpt/web-platform-tests/IndexedDB/idbobjectstore-query-exception-order.html

<!doctype html>
<meta charset=utf-8>
<title>IndexedDB: IDBObjectStore query method Ordering</title>
<link rel="help" href="https://w3c.github.io/IndexedDB/#dom-idbobjectstore-get">
<link rel="help" href="https://w3c.github.io/IndexedDB/#dom-idbobjectstore-getall">
<link rel="help" href="https://w3c.github.io/IndexedDB/#dom-idbobjectstore-getallkeys">
<link rel="help" href="https://w3c.github.io/IndexedDB/#dom-idbobjectstore-count">
<link rel="help" href="https://w3c.github.io/IndexedDB/#dom-idbobjectstore-opencursor">
<link rel="help" href="https://w3c.github.io/IndexedDB/#dom-idbobjectstore-openkeycursor">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support.js"></script>
<script>

['get',
 'getAll',
 'getAllKeys',
 'count',
 'openCursor',
 'openKeyCursor'
].forEach(method => {

  indexeddb_test(
    (t, db) => {
      const store = db.createObjectStore('s');
      const store2 = db.createObjectStore('s2');

      db.deleteObjectStore('s2');

      setTimeout(t.step_func(() => {
        assert_throws_dom(
          'InvalidStateError', () => { store2[method]('key'); },
          '"has been deleted" check (InvalidStateError) should precede ' +
          '"not active" check (TransactionInactiveError)');

        t.done();
      }), 0);
    },
    (t, db) => {},
    `IDBObjectStore.${method} exception order: ` +
    'InvalidStateError vs. TransactionInactiveError'
  );

  indexeddb_test(
    (t, db) => {
      const store = db.createObjectStore('s');
    },
    (t, db) => {
      const tx = db.transaction('s');
      const store = tx.objectStore('s');

      setTimeout(t.step_func(() => {
        assert_throws_dom(
          'TransactionInactiveError', () => { store[method]({}); },
          '"not active" check (TransactionInactiveError) should precede ' +
          'query check (DataError)');
        t.done();
      }), 0);
    },
    `IDBObjectStore.${method} exception order: ` +
    'TransactionInactiveError vs. DataError'
  );
});

</script>
